Rename environment -> environmentHelpers
